+++ changelog.md
@@ -26,6 +26,7 @@
   * Fishing range is now configurable
   * Pull strength is now capped
   * Increased fishing hook throw sound volume
+  * When hooked, the hook will now play a low-pitched splash sound 
 * Parkour is now disabled by default
 * Lowered default shulker armor
 * Enderman and wandering traders can no longer sprint